Rename QueryTermBase -> QueryTermUCS4 and move to separate h and cpp files.
