+++ searchcommon/src/vespa/searchcommon/attribute/i_search_context.h
@@ -7,7 +7,7 @@
 
 namespace search::fef { class TermFieldMatchData; }
 namespace search::queryeval { class SearchIterator; }
-namespace search { class QueryTermBase; }
+namespace search { class QueryTermUCS4; }
 
 namespace search::attribute {
 
@@ -47,7 +47,7 @@ public:
 
     virtual bool valid() const = 0;
     virtual Int64Range getAsIntegerTerm() const = 0;
-    virtual const QueryTermBase * queryTerm() const = 0;
+    virtual const QueryTermUCS4 * queryTerm() const = 0;
     virtual const vespalib::string &attributeName() const = 0;
 
     int32_t find(DocId docId, int32_t elementId, int32_t &weight) const { return onFind(docId, elementId, weight); }
