+++ searchlib/src/vespa/searchlib/attribute/multistringattribute.hpp
@@ -11,7 +11,7 @@
 #include <vespa/vespalib/text/lowercase.h>
 #include <vespa/vespalib/util/bufferwriter.h>
 #include <vespa/vespalib/util/regexp.h>
-#include <vespa/searchlib/query/queryterm.h>
+#include <vespa/searchlib/query/query_term_ucs4.h>
 
 namespace search {
