+++ searchlib/src/vespa/searchlib/query/CMakeLists.txt
@@ -2,6 +2,7 @@
 vespa_add_library(searchlib_query OBJECT
     SOURCES
     query_term_simple.cpp
+    query_term_ucs4.cpp
     queryterm.cpp
     querynode.cpp
     query.cpp
